+++ database/seeders/ClientSeeder.php
@@ -22,6 +22,7 @@ class ClientSeeder extends Seeder
                 'telephone' => null,
                 'type_client_id' => 2,
                 'zone_id' => 1,
+                'method_paid_id' => 1,
                 'comment' => 'Comentario sobre el cliente',
                 'active' => true,
                 'created_at' => now(),
@@ -32,6 +33,7 @@ class ClientSeeder extends Seeder
                 'telephone' => null,
                 'type_client_id' => 2,
                 'zone_id' => 2,
+                'method_paid_id' => 2,
                 'comment' => 'Comentario sobre el cliente',
                 'active' => true,
                 'created_at' => now(),
@@ -42,6 +44,7 @@ class ClientSeeder extends Seeder
                 'telephone' => null,
                 'type_client_id' => 2,
                 'zone_id' => 1,
+                'method_paid_id' => 1,
                 'comment' => 'Comentario sobre el cliente',
                 'active' => true,
                 'created_at' => now(),
@@ -52,6 +55,7 @@ class ClientSeeder extends Seeder
                 'telephone' => null,
                 'type_client_id' => 2,
                 'zone_id' => 1,
+                'method_paid_id' => 1,
                 'comment' => 'Comentario sobre el cliente',
                 'active' => true,
                 'created_at' => now(),
@@ -62,6 +66,7 @@ class ClientSeeder extends Seeder
                 'telephone' => null,
                 'type_client_id' => 2,
                 'zone_id' => 2,
+                'method_paid_id' => 3,
                 'comment' => 'Comentario sobre el cliente',
                 'active' => true,
                 'created_at' => now(),
